==> 01-02_sample_sp500_metadata.ipynb <==
# Sample - Metadata

NAME = '01-02_sample_sp500_metadata'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

import os
import re
import numpy as np
import pandas as pd

# Plotting
import matplotlib.pyplot as plt

# Logging
from utils import log_step

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

---
# Main Code

## Read metadata

METADATA_PATH = os.path.join('0_data', 'conference_calls_transcripts', 'metadata.csv')
metadata = pd.read_csv(METADATA_PATH, sep='\t', engine='python')
metadata.head()

## Pipeline

@log_step
def start_pipeline(metadata):
    return metadata.copy()

@log_step
def handle_dates(metadata):
    metadata['event_date'] = pd.to_datetime(metadata['event_date'], format=r'%d-%b-%Y')
    metadata['year'] = metadata['event_date'].dt.year
    return metadata

@log_step
def filter_sp500_firms(metadata):
    SP500_PATH = os.path.join('2_pipeline', '01-01_sample_sp500_constituents', 'out', 'sp500_consituents_2000-2020.feather')
    sp500 = pd.read_feather(SP500_PATH)

    metadata_sp500 = metadata.merge(sp500, on='gvkey')

    in_sp500 = metadata_sp500['event_date'].between(metadata_sp500['start_date'], metadata_sp500['end_date'])
    metadata_sp500 = metadata_sp500[in_sp500]
    
    return metadata_sp500

@log_step
def filter_earnings_calls(metadata):
    earning_dates = pd.read_csv(os.path.join('0_data', 'sp500_constituents', 'sp500_quarterly-earnings_2000-2020.csv'))
    
    earning_dates['event_date'] = pd.to_datetime(earning_dates['rdq'], format=r'%Y%m%d')
    earning_dates = earning_dates[['gvkey', 'event_date']].drop_duplicates()

    sp500_earnings = metadata.merge(earning_dates, on=['gvkey', 'event_date'], validate='m:1')

    # Drop duplicates if there are multiple conference calls on one day
    sp500_earnings = sp500_earnings.drop_duplicates(['gvkey', 'event_date']).reset_index()
    return sp500_earnings

def get_transcript_id_mapping(path):
    filepaths = []
    for year in os.listdir(path):
        year_path = os.path.join(path, year)
        if os.path.isdir(year_path):
            filenames = os.listdir(year_path)
            for filename in filenames:
                filepaths.append(os.path.join(year, filename))

    transcript_ids = [int(f.split('-')[-2]) for f in filepaths]
    return {k:v for k, v in zip(transcript_ids, filepaths)}

@log_step
def add_filepaths(metadata):
    TRANSCRIPTS_PATH = os.path.join('0_data', 'conference_calls_transcripts')
    transcript_id_mapping = get_transcript_id_mapping(TRANSCRIPTS_PATH)
    metadata['filepath'] = metadata['transcript_id'].apply(lambda x: transcript_id_mapping.get(x, np.nan))
    return metadata

@log_step
def clean_metadata(metadata):
    columns = ['gvkey', 'coname', 'event_date', 'event_name', 'event_time',
               'event_type', 'transcript_id', 'filepath', 'year',
               'cik', 'ticker', 'regcountrycode']
    
    metadata_clean = (metadata[columns]
                      .copy()
                      .rename(columns={'regcountrycode': 'country'})
                      .assign(gvkey=lambda x: x['gvkey'].astype('int'),
                              year=lambda x: x['year'].astype('int'))
                      .query('2004 <= year <= 2020')
                      .sort_values(['gvkey', 'event_date'])
                      .reset_index(drop=True))
    return metadata_clean

sample_metadata = (metadata.pipe(start_pipeline)
                           .pipe(handle_dates)
                           .pipe(filter_sp500_firms)
                           .pipe(filter_earnings_calls)
                           .pipe(add_filepaths)
                           .pipe(clean_metadata))
sample_metadata.head()

sample_metadata.to_feather(os.path.join(pipeline, 'out', 'sample_metadata.feather'))
