==> 03-02_model_finbert.ipynb <==
# Model - FinBERT

NAME = '03-02_model_finbert'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

import os
import re
import torch
from finbert_model import BertClassification

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

---
# Main Code 

PRETRAINED_WEIGHTS_PATH = os.path.join('0_data', 'finbert', 'finbert_pretrained_weights')
FINE_TUNED_WEIGHTS_PATH = os.path.join('0_data', 'finbert', 'huang_fine_tuned.pth')
DEVICE = 'cuda:0'

## Load FinBert Model

finbert = BertClassification(weight_path=PRETRAINED_WEIGHTS_PATH)

## Fine-tune FinBERT Model

finbert.load_state_dict(torch.load(FINE_TUNED_WEIGHTS_PATH, map_location=DEVICE))

## Save

torch.save(finbert, os.path.join(pipeline, 'out', 'finbert'))
