==> 04-01_tone_lm.ipynb <==
# Tone - LM Dictionary
*Source: [Loughran & McDonald (2011)](https://sraf.nd.edu/textual-analysis/resources/#Master%20Dictionary)*

NAME = '04-01_tone_lm'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

%%capture
import os
import re
import pickle
from datetime import datetime
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# Progress bar
from tqdm import tqdm
tqdm.pandas()

# NLP
from sklearn.feature_extraction.text import CountVectorizer

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

---
# Main Code 

cc = pd.read_feather(os.path.join('2_pipeline', '02-02_conference_calls_preprocess', 'out', 'cc_transcripts.feather'))
cc.head()

## Loughran & McDonald (2011)

with open(os.path.join('2_pipeline', '03-01_model_dictionaries', 'out', 'lm_positive.pickle'), 'rb') as f:
    lm_positive = pickle.load(f)

with open(os.path.join('2_pipeline', '03-01_model_dictionaries', 'out', 'lm_negative.pickle'), 'rb') as f:
    lm_negative = pickle.load(f)

print(f"# Positive: {len(lm_positive)}\n# Negative: {len(lm_negative)}")

lm_vocabulary = {j: i for i, j in enumerate(lm_positive + lm_negative)}
cv = CountVectorizer(vocabulary=lm_vocabulary)

cv_transcripts = cv.fit_transform(cc['transcript'])

# Vocabulary mapping
lm_positive_mapping = np.array([True if f in lm_positive else False for f in cv.get_feature_names()])
lm_negative_mapping = np.array([True if f in lm_negative else False for f in cv.get_feature_names()])

def sentiment_count(row):
    positive_count = cv_transcripts[row.name, lm_positive_mapping].sum()
    negative_count = cv_transcripts[row.name, lm_negative_mapping].sum()
    return positive_count, negative_count

cc['lm_positive'], cc['lm_negative'] = zip(*cc.progress_apply(sentiment_count, axis=1))

## Process Sentiment

def get_lm_tone(row):
    if (row['lm_positive'] + row['lm_negative']) > 0:  # avoid zero division error
        tone_lm = (row['lm_positive']-row['lm_negative']) / (row['lm_positive']+row['lm_negative'])
    else:
        tone_lm = 0
    return tone_lm

### Tone by Speaker

lm_tone_by_speaker = (cc
                      .groupby(['gvkey', 'ticker', 'event_date', 'speaker_role', 'speaker_name', 'speaker_firm'])
                      .agg(num_words=('num_words', np.sum),
                           num_sentences=('num_words', 'count'),
                           lm_positive=('lm_positive', np.sum),
                           lm_negative=('lm_negative', np.sum))
                      .assign(lm_tone=lambda x: x.apply(get_lm_tone, axis=1),
                              lm_tone_norm=lambda x: (x['lm_tone']-x['lm_tone'].mean()) 
                                          / x['lm_tone'].std())
                      .reset_index())
lm_tone_by_speaker

# Save
lm_tone_by_speaker.to_feather(os.path.join(pipeline, 'out', 'lm_tone_by_speaker.feather'))

### Tone by Role

lm_tone_by_role = (cc
                   .groupby(['gvkey', 'ticker', 'event_date', 'speaker_role'])
                   .agg(num_words=('num_words', np.sum),
                        num_sentences=('num_words', 'count'),
                        num_speakers=('speaker_name', 'nunique'),
                        lm_positive=('lm_positive', np.sum),
                        lm_negative=('lm_negative', np.sum))
                   .assign(lm_tone=lambda x: x.apply(get_lm_tone, axis=1),
                           lm_tone_norm=lambda x: (x['lm_tone']-x['lm_tone'].mean()) 
                         / x['lm_tone'].std())
                   .reset_index())

lm_tone_by_role

# Save
lm_tone_by_role.to_feather(os.path.join(pipeline, 'out', 'lm_tone_by_role.feather'))

### Sentiment by Firm

lm_tone_by_firm = (cc
                   .groupby(['gvkey', 'ticker', 'event_date'])
                   .agg(num_words=('num_words', np.sum),
                        num_sentences=('num_words', 'count'),
                        num_speakers=('speaker_name', 'nunique'),
                        lm_positive=('lm_positive', np.sum),
                        lm_negative=('lm_negative', np.sum))
                   .assign(lm_tone=lambda x: x.apply(get_lm_tone, axis=1),
                           lm_tone_norm=lambda x: (x['lm_tone']-x['lm_tone'].mean()) 
                                            / x['lm_tone'].std())
                   .reset_index())
lm_tone_by_firm

# Save
lm_tone_by_firm.to_feather(os.path.join(pipeline, 'out', 'lm_tone_by_firm.feather'))
