==> 04-02_tone_finbert.ipynb <==
# Tone - FinBERT

NAME = '04-02_tone_finbert'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

import os
import re
from tqdm import tqdm
import numpy as np
import pandas as pd
import torch
from torch.utils.data import TensorDataset, DataLoader
from finbert_utils import preprocess_for_finbert, finbert_predict

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

---
# Main Code 

CC_PATH = os.path.join('2_pipeline', '02-02_conference_calls_preprocess', 'out', 'cc_transcripts.feather')

cc = pd.read_feather(CC_PATH)
cc.head()

## Preprocess for BERT

LABELS = {0: 'neutral', 1 :'positive', 2: 'negative'}
NUM_LABELS = len(LABELS)
MAX_SEQ_LENGTH = 64
VOCAB_PATH = os.path.join('0_data', 'finbert', 'finbert_vocab')

input_ids, token_type_ids, attention_masks = preprocess_for_finbert(tqdm(cc['transcript']), VOCAB_PATH)

dataset = TensorDataset(input_ids, token_type_ids, attention_masks)
dataloader = DataLoader(dataset, batch_size=64)

## Predict with FinBERT

# Load FinBERT
finbert = torch.load(os.path.join('2_pipeline', '03-02_model_finbert', 'out', 'finbert'))

cc['finbert'] = finbert_predict(finbert, dataloader)

## Process FinBERT Tone

cc['finbert'].value_counts()

def get_finbert_tone(row):
    if (row['positive'] + row['negative']) > 0:  # avoid zero division error
        finbert_tone = (row['positive']-row['negative']) / (row['positive']+row['negative'])
    else:
        finbert_tone = 0
    return finbert_tone

### Tone by Speaker

finbert_tone_by_speaker = (cc
                           .groupby(['gvkey', 'ticker', 'event_date', 'speaker_role', 'speaker_name', 'speaker_firm', 'finbert'])
                           .size().unstack(fill_value=0)
                           .assign(finbert_tone=lambda x: x.apply(get_finbert_tone, axis=1),
                                   finbert_tone_norm=lambda x: (x['finbert_tone']-x['finbert_tone'].mean()) 
                                                                / x['finbert_tone'].std())
                           .reset_index())
finbert_tone_by_speaker

# Save
finbert_tone_by_speaker.to_feather(os.path.join(pipeline, 'out', 'finbert_tone_by_speaker.feather'))

### Tone by Role

finbert_tone_by_role = (cc
                        .groupby(['gvkey', 'ticker', 'event_date', 'speaker_role', 'finbert'])
                        .size().unstack(fill_value=0)
                        .assign(finbert_tone=lambda x: x.apply(get_finbert_tone, axis=1),
                                finbert_tone_norm=lambda x: (x['finbert_tone']-x['finbert_tone'].mean()) 
                                                        / x['finbert_tone'].std())
                        .reset_index())
finbert_tone_by_role

# Save
finbert_tone_by_role.to_feather(os.path.join(pipeline, 'out', 'finbert_tone_by_role.feather'))

### Tone by Firm

finbert_tone_by_firm = (cc
                        .groupby(['gvkey', 'ticker', 'event_date', 'finbert'])
                        .size().unstack(fill_value=0)
                        .assign(finbert_tone=lambda x: x.apply(get_finbert_tone, axis=1),
                                finbert_tone_norm=lambda x: (x['finbert_tone']-x['finbert_tone'].mean()) 
                                                        / x['finbert_tone'].std())
                        .reset_index())
finbert_tone_by_firm

# Save
finbert_tone_by_firm.to_feather(os.path.join(pipeline, 'out', 'finbert_tone_by_firm.feather'))
