==> 04-03_tone_measures.ipynb <==
# Tone - Combine

NAME = '04-03_tone_measures'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports

import os
import re
import numpy as np
import pandas as pd

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

---
# Main code

## Tone by Firm

lm_tone_by_firm = pd.read_feather(os.path.join('2_pipeline', '04-01_tone_lm', 'out', 'lm_tone_by_firm.feather'))
lm_tone_by_firm.head()

finbert_tone_by_firm = pd.read_feather(os.path.join('2_pipeline', '04-02_tone_finbert', 'out', 'finbert_tone_by_firm.feather'))
finbert_tone_by_firm.head()

tone_by_firm = (pd.merge(lm_tone_by_firm, finbert_tone_by_firm, how='outer', validate='1:1')
                .assign(year=lambda x: x['event_date'].dt.year,
                        quarter=lambda x: x['event_date'].dt.to_period('Q'))
                .filter(['gvkey', 'ticker', 'event_date', 'num_words', 'num_sentences',
                         'lm_tone', 'lm_tone_norm', 'finbert_tone', 'finbert_tone_norm',
                         'quarter', 'year']))
tone_by_firm

# Save
tone_by_firm.to_feather(os.path.join(pipeline, 'out', 'tone_by_firm.feather'))

## Tone by Role

lm_tone_by_role = pd.read_feather(os.path.join('2_pipeline', '04-01_tone_lm', 'out', 'lm_tone_by_role.feather'))
lm_tone_by_role.head()

finbert_tone_by_role = pd.read_feather(os.path.join('2_pipeline', '04-02_tone_finbert', 'out', 'finbert_tone_by_role.feather'))
finbert_tone_by_role.head()

tone_by_role = (pd.merge(lm_tone_by_role, finbert_tone_by_role, how='outer', validate='1:1')
                   .assign(year=lambda x: x['event_date'].dt.year,
                           quarter=lambda x: x['event_date'].dt.to_period('Q'))
                   .filter(['gvkey', 'ticker', 'event_date', 'speaker_role', 'num_words',
                            'num_sentences', 'lm_tone', 'lm_tone_norm', 'finbert_tone',
                            'finbert_tone_norm', 'quarter', 'year']))
tone_by_role

# Save
tone_by_role.to_feather(os.path.join(pipeline, 'out', 'tone_by_role.feather'))

## Tone by Speaker

lm_tone_by_speaker = pd.read_feather(os.path.join('2_pipeline', '04-01_tone_lm', 'out', 'lm_tone_by_speaker.feather'))
lm_tone_by_speaker.head()

finbert_tone_by_speaker = pd.read_feather(os.path.join('2_pipeline', '04-02_tone_finbert', 'out', 'finbert_tone_by_speaker.feather'))
finbert_tone_by_speaker.head()

tone_by_speaker = (pd.merge(lm_tone_by_speaker, finbert_tone_by_speaker, how='outer', validate='1:1')
                   .assign(year=lambda x: x['event_date'].dt.year,
                           quarter=lambda x: x['event_date'].dt.to_period('Q'))
                   .filter(['gvkey', 'ticker', 'event_date', 'speaker_role', 'speaker_name',
                            'speaker_firm', 'num_words', 'num_sentences', 'lm_tone', 'lm_tone_norm',
                            'finbert_tone', 'finbert_tone_norm', 'quarter', 'year']))
tone_by_speaker

# Save
tone_by_speaker.to_feather(os.path.join(pipeline, 'out', 'tone_by_speaker.feather'))
