==> 09-03_experiment_lm_accuracy.ipynb <==
# Experiment - LM Word Accuracy

NAME = '09-03_experiment_lm_accuracy'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports

import os
import re
import pickle
import numpy as np
import pandas as pd

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

---
# Main Code 
## Read Phrases
### Financial Phrasebank

def read_labeled_sentences(level):
    with open(os.path.join('0_data', 'financial_phrasebank', f'Sentences_{level}Agree.txt')) as f:
        labeled_sentences = f.readlines()
    sentences, labels = [], []
    for i in labeled_sentences:
        sentence, label = i.split(('@'))
        sentences.append(sentence)
        labels.append(label.strip())
    return pd.DataFrame({'sentence': sentences, 'label': labels})

fpb = read_labeled_sentences('50')
len(fpb)

### FiQa

NEUTRAL_THRESHOLD = 0.2

REPLACE_BY_SPACE = re.compile(r'[/(){}\[\]\|@,;-]')
BAD_SYMBOLS = re.compile(r'[^0-9a-z +_]')
HYPERLINKS = re.compile(r'(http)([^\s]+)')

def clean_text(transcript):
    transcript = transcript.lower()
    transcript = HYPERLINKS.sub('', transcript)
    transcript = REPLACE_BY_SPACE.sub(' ', transcript)
    transcript = BAD_SYMBOLS.sub('', transcript)
    transcript = re.sub(r'\s+', ' ', transcript)
    return transcript.strip()

headlines = pd.read_json(os.path.join('0_data', 'fiqa', 'task1_headline_ABSA_train.json')).transpose()

headlines['sentence'] = headlines['sentence'].apply(clean_text)

headlines['sentiment_score'] = (headlines['info']
                                .apply(lambda x: x[0]['sentiment_score'])
                                .astype(np.float))

headlines.loc[np.abs(headlines['sentiment_score']) <= NEUTRAL_THRESHOLD, 'label'] = 'neutral'
headlines.loc[headlines['sentiment_score'] > NEUTRAL_THRESHOLD, 'label'] = 'positive'
headlines.loc[headlines['sentiment_score'] < -NEUTRAL_THRESHOLD, 'label'] = 'negative'

headlines.drop(columns=['info'], inplace=True)

posts = pd.read_json(os.path.join('0_data', 'fiqa', 'task1_post_ABSA_train.json')).transpose()

posts['sentence'] = posts['sentence'].apply(clean_text)

posts['sentiment_score'] = (posts['info']
                            .apply(lambda x: x[0]['sentiment_score'])
                            .astype(np.float))

posts.loc[np.abs(posts['sentiment_score']) <= NEUTRAL_THRESHOLD, 'label'] = 'neutral'
posts.loc[posts['sentiment_score'] > NEUTRAL_THRESHOLD, 'label'] = 'positive'
posts.loc[posts['sentiment_score'] < -NEUTRAL_THRESHOLD, 'label'] = 'negative'

posts.drop(columns=['info'], inplace=True)

### Combine Phrases

phrases = pd.concat([fpb, headlines, posts])
len(phrases)

## Sentiment

### Loughran & McDonald (2011)

with open(os.path.join('2_pipeline', '03-01_model_dictionaries', 'out', 'lm_positive.pickle'), 'rb') as f:
    lm_positive = pickle.load(f)

with open(os.path.join('2_pipeline', '03-01_model_dictionaries', 'out', 'lm_negative.pickle'), 'rb') as f:
    lm_negative = pickle.load(f)

print(f"# Positive: {len(lm_positive)}\n# Negative: {len(lm_negative)}")

def count_from_list(sentence, word_list):
    '''Returns the count of words in word_list in conference calls'''
    sentence_words = sentence.lower().split()
    num = 0
    for word in word_list:
        num += sentence_words.count(word)
    return num

def sentiment_words(sentence, word_list):
    '''Returns the count of words in word_list in conference calls'''
    sentence_words = sentence.lower().split()
    words = []
    for word in sentence_words:
        if word in word_list:
            words.append(word)
    return words

def get_sentiment(row):
    '''Returns the sentiment classification'''
    if row['num_positive'] == row['num_negative']:
        return 'neutral'
    elif row['num_positive'] > row['num_negative']:
        return 'positive'
    else:
        return 'negative'

phrases['num_negative'] = phrases['sentence'].apply(count_from_list, word_list=lm_negative)
phrases['num_positive'] = phrases['sentence'].apply(count_from_list, word_list=lm_positive)

phrases['words'] = phrases['sentence'].apply(sentiment_words, word_list=lm_positive + lm_negative)

phrases['lm_dictionary'] = phrases.apply(get_sentiment, axis=1)

sentiment_words = (phrases
                   .assign(word=lambda x: x['words'])
                   .explode('word')
                   .drop_duplicates(['sentence', 'word']))
sentiment_words

word_overview = (sentiment_words
                 .loc[sentiment_words['words'].str.len() == 1]
                 ['word'].value_counts()
                 .reset_index()
                 .rename(columns={'index': 'word', 'word': 'frequency'})
                 .assign(lm_sentiment=lambda x: x['word'].apply(lambda x: 'Positive' if x in lm_positive else 'Negative')))

word_overview

def get_actual_sentiment(word):
    return (sentiment_words
            .loc[sentiment_words['words'].str.len() == 1]
            .loc[lambda x: x['word'] == word]
            .drop_duplicates('sentence')
            ['label'].value_counts())

word_overview[['Negative', 'Positive', 'Neutral']] = word_overview['word'].apply(lambda x: get_actual_sentiment(x)).fillna(0).astype('int')

word_overview['num_correct'] = word_overview.apply(lambda x: x[x['lm_sentiment']], axis=1)
word_overview['Accuracy'] = word_overview['num_correct'] / word_overview['frequency']

## Tables

(word_overview
.query("frequency >= 5 & lm_sentiment == 'Positive'")
.sort_values('Accuracy', ascending=False)
.head(8)
.set_index('word')
.rename(columns={'frequency': 'Frequency',
                 'lm_sentiment': 'LM Dictionary'})
.assign(Accuracy=lambda x: x['Accuracy'].apply(lambda s: f'{s:.2%}'))
.filter(['Frequency', 'Positive',
         'Neutral', 'Negative', 'Accuracy']))

(word_overview
.query("frequency >= 5 & lm_sentiment == 'Positive'")
.sort_values('Accuracy', ascending=False)
.tail(8)
.set_index('word')
.rename(columns={'frequency': 'Frequency',
                 'lm_sentiment': 'LM Dictionary'})
.assign(Accuracy=lambda x: x['Accuracy'].apply(lambda s: f'{s:.2%}'))
.filter(['Frequency', 'Positive',
         'Neutral', 'Negative', 'Accuracy']))

(word_overview
.query("frequency >= 5 & lm_sentiment == 'Negative'")
.sort_values('Accuracy', ascending=False)
.head(8)
.set_index('word')
.rename(columns={'frequency': 'Frequency',
                 'lm_sentiment': 'LM Dictionary'})
.assign(Accuracy=lambda x: x['Accuracy'].apply(lambda s: f'{s:.2%}'))
.filter(['Frequency', 'Positive',
         'Neutral', 'Negative', 'Accuracy']))

(word_overview
.query("frequency >= 5 & lm_sentiment == 'Negative'")
.sort_values('Accuracy', ascending=False)
.tail(8)
.set_index('word')
.rename(columns={'frequency': 'Frequency',
                 'lm_sentiment': 'LM Dictionary'})
.assign(Accuracy=lambda x: x['Accuracy'].apply(lambda s: f'{s:.2%}'))
.filter(['Frequency', 'Positive',
         'Neutral', 'Negative', 'Accuracy']))
