==> 09-04_experiment_lm_did.ipynb <==
# Experiment - Loughran & McDonald Difference-in-difference

NAME = '09-04_experiment_lm_did'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

import os
import re
import numpy as np
import pandas as pd

# Regressions
from linearmodels import PanelOLS
import statsmodels.api as sm

# Hide annoying warnings
import warnings
warnings.filterwarnings('ignore')

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

pipeline = os.path.join('2_pipeline', NAME)
if not os.path.exists(pipeline):
    os.makedirs(pipeline)
    for folder in ['out', 'store', 'tmp']:
        os.makedirs(os.path.join(pipeline, folder))

--- 
# Main code

# Preprocess Transcripts

controls = pd.read_feather(os.path.join('2_pipeline', '07-01_dataset_control_variables', 'out', 'control_variables.feather'))
controls.head()

LM_PUBLICATION = '2011-01-06'  # Date of Loughran & McDonald (2011) Publication
tone_by_role = pd.read_feather(os.path.join('2_pipeline', '04-01_tone_lm', 'out', 'lm_tone_by_role.feather'))

tone_did = (tone_by_role
            .assign(year=lambda x: x['event_date'].dt.year,
                    quarter=lambda x: x['event_date'].dt.to_period('Q'),
                    treated=lambda x: np.where(x['speaker_role'] == 'Management', 1, 0),
                    post=lambda x: (x['event_date'] >= LM_PUBLICATION).astype('int'),
                    treated_post=lambda x: x['post'] * x['treated'])
            .merge(controls, on=['gvkey', 'quarter'])
            .assign(quarter=lambda x: x['quarter'].astype('str'))  # Stata cannot handle the pandas.Period dtype
            .dropna()
            .set_index(['gvkey', 'year']))

tone_did

tone_did['treated'].value_counts()

# Save
tone_did.to_stata(os.path.join(pipeline, 'out', 'tone_did.dta'))

## Diffrence-in-difference

controls = ['surprise', 'size', 'roa', 'leverage', 'loss', 'capex']

def ols_model(df, dependent, exog, ffe, tfe):
    dependent = df[dependent]
    exog = sm.add_constant(df[exog])
    model = PanelOLS(dependent, exog, entity_effects=ffe, time_effects=tfe)
    return model.fit(cov_type='clustered', cluster_entity=True)

### Negative Words

ols_model(df=tone_did, 
          dependent='lm_negative',
          exog=['treated', 'post', 'treated_post' ,'num_words'],
          ffe=False,
          tfe=False)

ols_model(df=tone_did, 
          dependent='lm_negative',
          exog=['treated', 'post', 'treated_post', 'num_words'] + controls,
          ffe=False,
          tfe=False)

ols_model(df=tone_did, 
          dependent='lm_negative',
          exog=['treated', 'post', 'treated_post', 'num_words'],
          ffe=False,
          tfe=True)

ols_model(df=tone_did, 
          dependent='lm_negative',
          exog=['treated', 'post', 'treated_post', 'num_words'],
          ffe=True,
          tfe=True)

ols_model(df=tone_did, 
          dependent='lm_positive',
          exog=['treated', 'post', 'treated_post', 'num_words'],
          ffe=False,
          tfe=False)

ols_model(df=tone_did, 
          dependent='lm_positive',
          exog=['treated', 'post', 'treated_post', 'num_words'] + controls,
          ffe=False,
          tfe=False)

ols_model(df=tone_did, 
          dependent='lm_positive',
          exog=['treated', 'post', 'treated_post', 'num_words'],
          ffe=False,
          tfe=True)

ols_model(df=tone_did, 
          dependent='lm_positive',
          exog=['treated', 'post', 'treated_post', 'num_words'],
          ffe=True,
          tfe=True)
