==> 09-06_experiment_analysts.ipynb <==
# Experiment - Analyst Output

NAME = '09-06-experiment_analysts'
PROJECT = 'conference-calls-sentiment'
PYTHON_VERSION = '3.7.0'

### Imports  

import os
import re
import numpy as np
import pandas as pd

# Regressions
from linearmodels import PanelOLS
import statsmodels.api as sm

# Hide annoying warnings
import warnings
warnings.filterwarnings('ignore')

### Settings

workdir = re.sub("(?<={})[\w\W]*".format(PROJECT), "", os.getcwd())
os.chdir(workdir)

--- 
# Main code

analysts_dataset = pd.read_feather(os.path.join('2_pipeline', '07-02_dataset_construction', 'out', 'analysts_dataset.feather'))
analysts_dataset = analysts_dataset.set_index(['gvkey', 'year'])
analysts_dataset.head()

def ols_model(df, tone, dependent):
    dependent = df[dependent]
    exog_vars = [tone, 'num_words', 'surprise', 'size', 'roa', 'leverage', 'loss', 'capex']
    exog = sm.add_constant(df[exog_vars])
    model = PanelOLS(dependent, exog, entity_effects=True, time_effects=True)
    return model.fit(cov_type='clustered', cluster_entity=True)

## Regressions

def get_stars(pvalue):
    if pvalue < 0.01:
        return '***'
    if pvalue < 0.05:
        return '**'
    if pvalue < 0.1:
        return '*'
    else:
        return ''

for dv in ['rating_change', 'price_change', 'eps_change']:
    print(dv.upper())
    print('=' * 50)
    for tone in ['lm_tone_norm', 'finbert_tone_norm']:
        res = ols_model(analysts_dataset, tone, dv)
        for name, coefficient, tstat, pvalue in zip(res._var_names, res.params, res.tstats, res.pvalues):
            print(f"{name:18s} {coefficient:10.3f}{get_stars(pvalue):3s}\t({tstat:.2f})")
        print('-' * 50)
    print('')

## Bootstrapping

def create_sample(df, dependent, n_obs):
    sample = (df
              .copy()
              .filter([dv, 'lm_tone_norm', 'finbert_tone_norm', 'num_words',
                       'surprise', 'size', 'roa', 'leverage', 'loss', 'capex'])
              .dropna()
              .sample(n_obs))
    return sample

N_ITERS = 500
N_OBS = 2500

columns = ['dv', 'tone', 'coefficient', 'pvalue']
results = pd.DataFrame(columns=columns)
for dv in ['rating_change', 'price_change', 'eps_change']:
    for tone in ['lm_tone_norm', 'finbert_tone_norm']:
        for _ in range(N_ITERS):
            # Create random sample
            sample = create_sample(analysts_dataset, dv, N_OBS)
            res = ols_model(sample, tone, dv)

            # Get coefficient and p-value
            coefficient, pvalue = res.params[tone], res.pvalues[tone]

            # Append to results
            results.loc[len(results)] = [dv, tone, coefficient, pvalue]

table = (results.assign(positive_sign=lambda x: (x['coefficient'] > 0),
                        pct_10=lambda x: x['positive_sign'] & (x['pvalue'] < 0.10),
                        pct_5=lambda x: x['positive_sign'] & (x['pvalue'] < 0.05),
                        pct_1=lambda x: x['positive_sign'] & (x['pvalue'] < 0.01))
                .groupby(['dv', 'tone'])
                .agg({'coefficient': np.mean,
                      'positive_sign': np.mean,
                      'pct_10': np.mean,
                      'pct_5': np.mean,
                      'pct_1': np.mean})
                .sort_index(ascending=False)
                .round(3))
table

print(table.to_latex())
